test: move test to correct file
